# docs/swagger.yaml
basePath: /
definitions:
...
  domain.CreateRoleRequest:
    properties:
      abilities:
        items:
          type: string
        type: array
      description:
        type: string
      max_count:
        type: integer
      name:
        type: string
      team:
        type: string
    required:
    - name
    type: object
...
  domain.GameRule:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      id:
        type: integer
      name:
        type: string
      phase:
        type: string
      updated_at:
        type: string
    type: object
...
  domain.Role:
    properties:
      abilities:
        items:
          type: string
        type: array
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      max_count:
        type: integer
      name:
        type: string
      team:
        type: string
      updated_at:
        type: string
    type: object
  domain.RuleRequest:
    properties:
      description:
        type: string
      enabled:
        type: boolean
      name:
        type: string
      phase:
        type: string
    required:
    - name
    type: object
  domain.Scenario:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      roles:
        items:
          type: string
        type: array
      rules:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
  domain.ScenarioRequest:
    properties:
      description:
        type: string
      name:
        type: string
      roles:
        items:
          type: string
        type: array
      rules:
        items:
          type: string
        type: array
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the Mafia game.
  title: Mafia API
  version: "1.0"
paths:
  /admin/abilities:
    get:
      description: Returns all predefined abilities for roles.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - BearerAuth: []
      summary: List available abilities
      tags:
      - Admin
  /admin/roles:
    get:
      description: Lists all configured roles.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Role'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List roles
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Creates a new role definition.
      parameters:
      - description: Role payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a role
      tags:
      - Admin
  /admin/roles/{id}:
    delete:
      description: Deletes a role by ID.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a role
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Updates an existing role definition.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a role
      tags:
      - Admin
  /admin/rules:
    post:
      consumes:
      - application/json
      description: Creates a new game rule.
      parameters:
      - description: Rule payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.GameRule'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a rule
      tags:
      - Admin
  /admin/rules/{id}:
    delete:
      description: Deletes a game rule by ID.
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a rule
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Updates an existing game rule by ID.
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: integer
      - description: Rule payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GameRule'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a rule
      tags:
      - Admin
  /admin/scenarios:
    get:
      description: Lists all configured scenarios.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Scenario'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List scenarios
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Creates a new scenario with rules and roles.
      parameters:
      - description: Scenario payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ScenarioRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Scenario'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a scenario
      tags:
      - Admin
  /admin/scenarios/{id}:
    delete:
      description: Deletes a scenario by ID.
      parameters:
      - description: Scenario ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a scenario
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Updates an existing scenario by ID.
      parameters:
      - description: Scenario ID
        in: path
        name: id
        required: true
        type: integer
      - description: Scenario payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ScenarioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Scenario'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a scenario
      tags:
      - Admin
...
swagger: "2.0"